+++ Source/Graphics/PDFGraphics.swift
@@ -15,7 +15,7 @@
 
  - Image resizing and compression
  */
-class PDFGraphics {
+internal enum PDFGraphics {
 
     // MARK: - INTERNAL STATIC FUNCS
 
@@ -28,7 +28,7 @@ class PDFGraphics {
      - parameter end: End point of line
      - parameter style: Style of drawn line
      */
-    static func drawLine(start: CGPoint, end: CGPoint, style: PDFLineStyle) {
+    internal static func drawLine(start: CGPoint, end: CGPoint, style: PDFLineStyle) {
         if let path = createLinePath(start: start, end: end, style: style) {
             style.color.setStroke()
 
@@ -45,7 +45,7 @@ class PDFGraphics {
 
      - returns: Bezier path of line, `nil` if line type in `style` was `PDFLineType.none`
      */
-    static func createLinePath(start: CGPoint, end: CGPoint, style: PDFLineStyle) -> UIBezierPath? {
+    internal static func createLinePath(start: CGPoint, end: CGPoint, style: PDFLineStyle) -> UIBezierPath? {
         if style.type == .none {
             return nil
         }
@@ -70,7 +70,7 @@ class PDFGraphics {
      - parameter outline: Style of border lines
      - parameter fill: Inner color
      */
-    static func drawRect(rect: CGRect, outline: PDFLineStyle, fill: UIColor = .clear) {
+    internal static func drawRect(rect: CGRect, outline: PDFLineStyle, fill: UIColor = .clear) {
         let path = createRectPath(rect: rect, outline: outline)
 
         outline.color.setStroke()
@@ -87,7 +87,7 @@ class PDFGraphics {
      - parameter outline: Style of border lines
      - parameter fill: Inner color
      */
-    static func drawRect(rect: CGRect, outline: PDFLineStyle, pattern: FillPattern) {
+    internal static func drawRect(rect: CGRect, outline: PDFLineStyle, pattern: FillPattern) {
         let path = createRectPath(rect: rect, outline: outline)
 
         outline.color.setStroke()
@@ -103,7 +103,7 @@ class PDFGraphics {
      - parameter rect: Frame of rectangle
      - parameter outline: Style of border lines
      */
-    static func createRectPath(rect: CGRect, outline: PDFLineStyle) -> UIBezierPath {
+    internal static func createRectPath(rect: CGRect, outline: PDFLineStyle) -> UIBezierPath {
         var path = UIBezierPath(rect: rect)
         if let radius = outline.radius {
             path = UIBezierPath.init(roundedRect: rect, cornerRadius: radius)
@@ -127,7 +127,7 @@ class PDFGraphics {
 
      - returns: Array with dash values
      */
-    static func createDashes(style: PDFLineStyle, path: inout UIBezierPath) -> [CGFloat] {
+    internal static func createDashes(style: PDFLineStyle, path: inout UIBezierPath) -> [CGFloat] {
         var dashes: [CGFloat] = []
 
         switch style.type {
@@ -144,18 +144,43 @@ class PDFGraphics {
         return dashes
     }
 
+    /**
+     TODO: Documentation
+     */
+    internal static func drawPath(path: UIBezierPath, outline: PDFLineStyle, fillColor: UIColor) {
+        guard var path = path.copy() as? UIBezierPath else {
+            fatalError("Copy of UIBezierPath is invalid!")
+        }
+        let dashes = createDashes(style: outline, path: &path)
+        path.setLineDash(dashes, count: dashes.count, phase: 0.0)
+        path.lineWidth = CGFloat(outline.width)
+
+        outline.color.setStroke()
+        fillColor.setFill()
+
+        path.fill()
+        path.stroke()
+    }
+
     // MARK: - Image Manipulation
 
     /**
-     Draws a scaled version of the given `image` into the given `frame`.
+     Creates a scaled version of the given `image` by the given `frame` and add optional corner clipping.
 
      - parameter image: Image to resize and compress
      - parameter frame: Frame in which the new image will fit
+     - parameter shouldResize: Flag if image should be resized to frame before drawing
+     - parameter shouldCompress: Flag if image should be compressed to quality before drawing
      - parameter quality: Value between 0.0 and 1.0, where 1.0 is maximum quality
+     - parameter roundCorners: Indicates which corners should be rounded, defaults to be empty
+     - parameter cornerRadii: Radius of corners which are rounded based on `roundCorners`, if `nil` it will create a fully round image
 
-     - returns: Resized and compressed version of `image`
+     - returns: Resized, compressed and rounded copy of `image`
      */
-    static func resizeAndCompressImage(image: UIImage, frame: CGRect, shouldResize: Bool, shouldCompress: Bool, quality: CGFloat) -> UIImage {
+    internal static func resizeAndCompressImage(image: UIImage, frame: CGRect,
+                                                shouldResize: Bool, shouldCompress: Bool,
+                                                quality: CGFloat,
+                                                roundCorners: UIRectCorner = [], cornerRadius: CGFloat? = nil) -> UIImage {
         var finalImage = image
 
         if shouldResize {
@@ -163,6 +188,9 @@ class PDFGraphics {
         }
         if shouldCompress {
             finalImage = compress(image: finalImage, quality: quality)
+        }
+        if !roundCorners.isEmpty {
+            finalImage = round(image: finalImage, frameSize: frame.size, corners: roundCorners, cornerRadius: cornerRadius)
         }
 
         return finalImage
@@ -177,7 +205,7 @@ class PDFGraphics {
 
      - returns: Resized version of `image`
      */
-    static func resize(image: UIImage, frame: CGRect, quality: CGFloat) -> UIImage {
+    internal static func resize(image: UIImage, frame: CGRect, quality: CGFloat) -> UIImage {
         let factor: CGFloat = min(3 * quality, 1)
         let resizeFactor = factor.isZero ? 0.2 : factor
 
@@ -198,8 +226,10 @@ class PDFGraphics {
 
      - parameter image: Image to compress
      - parameter quality: Value between 0.0 and 1.0, where 1.0 is maximum quality/least compression
+
+     - returns: Compressed image
      */
-    static func compress(image: UIImage, quality: CGFloat) -> UIImage {
+    internal static func compress(image: UIImage, quality: CGFloat) -> UIImage {
         guard let data = image.jpegData(compressionQuality: quality) else {
             return image
         }
@@ -210,13 +240,52 @@ class PDFGraphics {
     }
 
     /**
+     Clips an image to add round corners
+
+     - parameter image: Image to edit
+     - parameter size: Size of final image, used to calculate correct radii limits
+     - parameter cornerRadius: Optional value used as radius, if null half of the minimum of the size width or height is used,
+     resulting in a round image
+
+     - returns: Manipulated image
+     */
+    internal static func round(image: UIImage, frameSize: CGSize, corners: UIRectCorner, cornerRadius: CGFloat?) -> UIImage {
+        let size = image.size
+
+        var cornerRadii = CGSize.zero
+        if var radius = cornerRadius {
+            radius = radius * (frameSize.width > frameSize.height ? size.width / frameSize.width : size.height / frameSize.height)
+            cornerRadii = CGSize(width: radius, height: radius)
+        } else {
+            let radius = size.width < size.height ? frameSize.width / 2 : frameSize.height / 2
+            cornerRadii = CGSize(width: radius, height: radius)
+        }
+
+        UIGraphicsBeginImageContext(size)
+        let clipPath = UIBezierPath(roundedRect: CGRect(origin: .zero, size: size), byRoundingCorners: corners, cornerRadii: cornerRadii)
+        clipPath.addClip()
+        image.draw(in: CGRect(origin: .zero, size: size))
+
+        let finalImage = UIGraphicsGetImageFromCurrentImageContext()
+        UIGraphicsEndImageContext()
+
+        return finalImage ?? image
+    }
+
+    /**
      Constants for filling, mostly used for debugging elements
      */
-    enum FillPattern {
+    internal enum FillPattern {
 
+        /**
+         TODO: Documentation
+         */
         case dotted(foreColor: UIColor, backColor: UIColor)
 
-        func setFill() {
+        /**
+         TODO: Documentation
+         */
+        internal func setFill() {
             switch self {
             case .dotted(let foreColor, let backColor):
                 UIGraphicsBeginImageContext(CGSize(width: 5, height: 5))
@@ -236,7 +305,5 @@ class PDFGraphics {
                 UIColor(patternImage: image!).setFill()
             }
         }
-
     }
-
 }